Python/ShowAIAuto-main/requirements.txt
python-dotenv
requests
playwright
agentql 